Move expression diagnostics regression tests to xUnit
